+++ src/shelly_analyzer/__init__.py
@@ -2,4 +2,4 @@
 
 __all__ = ["__version__"]
 
-__version__ = "13.2.3"
+__version__ = "13.3.0"